fix bug in initial conditions

+++ Examples/MAVMono/Online/src/MAVCom/location_manager.h
@@ -21,20 +21,20 @@ public:
     void set_initial_geodetic_pose();
     void get_NED_from_geodetic(double lat, double lon, double alt, float *x, float *y, float *z);
 
-    void set_local_position(uint64_t timestamp, double x, double y, double z);
-    void set_global_position(uint64_t timestamp, double lat, double lon, double alt);
+    void set_local_position(uint32_t timestamp, double x, double y, double z);
+    void set_global_position(uint32_t timestamp, double lat, double lon, double alt);
     bool isGeodeticInitialize();
 
     double distanceInKmBetweenEarthCoordinates(double lat1, double lon1, double lat2, double lon2);
 
-    double interpolate(uint64_t x1, uint64_t x2, uint64_t x_predict, double y1, double y2);
+    double interpolate(uint32_t x1, uint32_t x2, uint32_t x_predict, double y1, double y2);
 
 private:
     geodetic_converter::GeodeticConverter *geodeticConverter;
 
     double cx,cy,cz;
     double c_lat, c_lon, c_alt;
-    uint64_t c_local_timestamp, c_global_timestamp;
+    uint32_t c_local_timestamp, c_global_timestamp;
 
     pthread_mutex_t mutex_localpose, mutex_globalpose;
 };
